requirements.txt
aspy.yaml==1.3.0
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
cffi==1.13.1
cfgv==2.0.1
cryptography==2.8
future==0.18.1
identify==1.4.7
more-itertools==7.2.0
nodeenv==1.3.3
packaging==19.2
pipenv==2018.11.26
pluggy==0.13.0
pre-commit==1.20.0
py==1.8.0
pycparser==2.19
pyparsing==2.4.2
pytest==5.2.2
python-dotenv==0.10.3
python-telegram-bot==12.2.0
PyYAML==5.1.2
six==1.12.0
toml==0.10.0
tornado==6.0.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
wcwidth==0.1.7
